+++ table/block_based/block_based_table_reader_test.cc
@@ -67,8 +67,8 @@ class BlockBasedTableReaderTest
     std::unique_ptr<TableBuilder> table_builder(table_factory_->NewTableBuilder(
         TableBuilderOptions(ioptions, moptions, comparator, &factories,
                             compression_type, CompressionOptions(),
-                            false /* skip_filters */, 0 /* column_family_id */,
-                            kDefaultColumnFamilyName, -1 /* level */),
+                            0 /* column_family_id */, kDefaultColumnFamilyName,
+                            -1 /* level */),
         writer.get()));
 
     // Build table.