+++ table/block_based/data_block_hash_index_test.cc
@@ -558,7 +558,7 @@ void TestBoundary(InternalKey& ik1, std::string& v1, InternalKey& ik2,
       TableBuilderOptions(
           ioptions, moptions, internal_comparator,
           &int_tbl_prop_collector_factories, options.compression,
-          CompressionOptions(), false /* skip_filters */,
+          CompressionOptions(),
           TablePropertiesCollectorFactory::Context::kUnknownColumnFamily,
           column_family_name, level_),
       file_writer.get()));